+++ dsa_c/src/2_linked_lists.h
@@ -4,7 +4,7 @@
 // --- Singly Linked List (SLL) ---
 
 struct SLLNode {
-    int data;
+    long data;
     struct SLLNode* next;
 };
 
@@ -15,8 +15,8 @@ struct SLL {
 // SLL function prototypes
 struct SLL* sll_create();
 void sll_destroy(struct SLL* list);
-struct SLLNode* sll_createNode(int data);
-void sll_insert(struct SLL* list, int index, int new_data);
+struct SLLNode* sll_createNode(long data);
+void sll_insert(struct SLL* list, int index, long new_data);
 int sll_length(struct SLL* list);
 void sll_delete(struct SLL* list, int position);
 void sll_print(struct SLL* list);
